+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [0.1.1] - 2026-03-17
 - Fixed: CSS class not applied to non-recurring events in reader
+- Fixed: backend asset loading
 
 ## [0.1.0] - 2026-03-14
 Initial release